Crate quicksort

A rabbit hole to quicksort

ALGORITHM 63

PARTITION

C. A. R. HOARE

Elliott Brothers Ltd., Borehamwood, Hertfordshire, Eng.

procedure	partition (A,M,N,I,J);	value M,N;
		array A;	integer	 M,N,I,J;

comment I and J are output variables, and A is the array (with subscript bounds M:N) which is operated by this procdeure. Partition take the value X of a random element of the array A, and rearrange the values of the elements of the arrary in such a way that there exists integers I and J with the following properties:

M	<=	J	<	I	 <=	N	provided M < N
A[R]	<=	X	for M <= R <= J
A[R]	==	X	for J <  R  < I
A[R]	>=	X	for I <= R <= N

The procedure uses an integer procedure random (M,N) which chooses equiprobably a random interger F between M and N, and also a procedure exchange, which exchanges the value of its two parameters;

begin	real X; integer F;
	F := random (M,N); X := A[F];
	I := M; J := N;
up:	for I := I	step 1	until N do
		if X < A[I]	then	go to	down;
	I := N;
down:	for J := J	step -1	until M do
		if A[J] < X	then 	go to	change;
	J := M;
change:	for I < J	then	begin	exchange (A[I], A[J]);
				I := I + 1; J := J - 1;			
				go to	up
			end
else	if I < F then	begin	exchange (A[I], A[F]);
				I := I + 1
			end
else	if F< J	then	begin	exchange (A[F], A[J]);
				J := J - 1
			end;
end    partition

ALGORITHM 64

QUICKSORT

C. A. R. HOARE

Elliott Brothers Ltd., Borehamwood, Hertfordshire, Eng.

procedure	quicksort (A,M,N);	value M,N;
		array A;	integer M,N;

comment Quicksort is a very fast and convenient method of sorting an array in the random-access store of a computer. The entire contents of the store may be sorted, since no extra space is required. The average number of comparisons made is 2(M-N) ln(N-M), and the average number of exchanges is one sixth this amount. Suitable refinements of this method will be desirable for its implementation on any actual computer.

begin	integer I,J;
	if M < N	then	begin	partition (A,M,N,I,J);      
					quicksort(A,M,J);
					quicksort(A,I,N)
				end
end	quicksort

Implementation in Rust

To implement quicksort, we need rand

[package]
name = "quicksort"
version = "0.1.0"
edition = "2024"

[dependencies]
rand = "0.10.0"

Functions

main ๐