Crate quicksort

A rabbit hole to quicksort

ALGORITHM 63

PARTITION

C. A. R. HOARE

Elliott Brothers Ltd., Borehamwood, Hertfordshire, Eng.

procedure	partition (A,M,N,I,J);	value M,N;
		array A;	integer	 M,N,I,J;

comment I and J are output variables, and A is the array (with subscript bounds M:N) which is operated by this procdeure. Partition take the value X of a random element of the array A, and rearrange the values of the elements of the arrary in such a way that there exists integers I and J with the following properties:

M	<=	J	<	I	 <=	N	provided M < N
A[R]	<=	X	for M <= R <= J
A[R]	==	X	for J <  R  < I
A[R]	>=	X	for I <= R <= N

The procedure uses an integer procedure random (M,N) which chooses equiprobably a random interger F between M and N, and also a procedure exchange, which exchanges the value of its two parameters;

begin	real X; integer F;
	F := random (M,N); X := A[F];
	I := M; J := N;
up:	for I := I	step 1	until N do
		if X < A[I]	then	go to	down;
	I := N;
down:	for J := J	step -1	until M do
		if A[J] < X	then 	go to	change;
	J := M;
change:	if I < J	then	begin	exchange (A[I], A[J]);
				I := I + 1; J := J - 1;			
				go to	up
			end
else	if I < F then	begin	exchange (A[I], A[F]);
				I := I + 1
			end
else	if F< J	then	begin	exchange (A[F], A[J]);
				J := J - 1
			end;
end    partition

ALGORITHM 64

QUICKSORT

C. A. R. HOARE

Elliott Brothers Ltd., Borehamwood, Hertfordshire, Eng.

procedure	quicksort (A,M,N);	value M,N;
		array A;	integer M,N;

comment Quicksort is a very fast and convenient method of sorting an array in the random-access store of a computer. The entire contents of the store may be sorted, since no extra space is required. The average number of comparisons made is 2(M-N) ln(N-M), and the average number of exchanges is one sixth this amount. Suitable refinements of this method will be desirable for its implementation on any actual computer.

begin	integer I,J;
	if M < N	then	begin	partition (A,M,N,I,J);      
					quicksort(A,M,J);
					quicksort(A,I,N)
				end
end	quicksort

ALGORITHM 65

FIND

C. A. R. HOARE

Elliott Brothers Ltd., Borehamwood, Hertfordshire, Eng.

procedure	find (A,M,N,K);	value M,N,K;
		array A;	integer M,N,K;

comment Find will assign to A[K] the value which it would behave if the array A[M:N] had been sorted. The array A will be partly sorted, and subsequent entries will be faster than the first;

begin	integer I,J;
	if M < N	then	begin	partition (A,M,N,I,J);      
					if K <= I then find(A,M,J,K);
					else if J <= K then find(A,J,N,K)
				end
end	find

https://doi.org/10.1145/366622.366642

PARTITION by hand

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I       F       J

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I       F     J 

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I       F     J 

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I       F   J    

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I       F   J    

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I       F J      

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I       F       
        J

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I     J F       

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I   J   F       

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I   J   F       

M               N
1 2 3 4 5 6 7 8 9
P A R T I T I O N
I J     F       

M               N
1 2 3 4 5 6 7 8 9
A P R T I T I O N
I J     F       

M               N
1 2 3 4 5 6 7 8 9
A P R T I T I O N
J I     F       

M               N
1 2 3 4 5 6 7 8 9
A I R T P T I O N
J I     F       

M               N
1 2 3 4 5 6 7 8 9
A I R T P T I O N
J   I   F       

Confirm it satistfies property

M	<=	J	<	I	 <=	N	provided M < N
A[R]	<=	X	for M <= R <= J
A[R]	==	X	for J <  R  < I
A[R]	>=	X	for I <= R <= N

1	<=	1	<	3	 <=	9	provided M < N
A[R]	<=	I	for 1 <= R <= 1
A[R]	==	I	for 1 <  R  < 3
A[R]	>=	I	for 3 <= R <= 9

Implementation in C (generated by AI)

#include <stdio.h>
#include <stdlib.h>
#include <string.h>

void exchange(char *a, char *b) {
    char temp = *a;
    *a = *b;
    *b = temp;
}

void partition(char A[], int M, int N, int *final_I, int *final_J) {

    int F = M + rand() % (N - M + 1);
    char X = A[F];

    int I = M;
    int J = N;

up:
    for (; I <= N; I++) {
        if (X < A[I]) goto down;
    }
    I = N;

down:
    for (; J >= M; J--) {
        if (A[J] < X) goto change;
    }
    J = M;

change:
    if (I < J) {
       exchange(&A[I], &A[J]); 
       I = I + 1;
       J = J - 1;
       goto up;
    }
    else if (I < F) {
       exchange(&A[I], &A[F]); 
       I = I + 1;
    }
    else if (F < J) {
       exchange(&A[F], &A[J]); 
       J = J - 1;
    }
    *final_I = I;
    *final_J = J;
}

void quicksort(char A[], int M, int N) {
    int I;
    int J;
    if (M < N) {
        partition(A, M, N, &I, &J);
        quicksort(A, M, J); 
        quicksort(A, I, N); 
    }
}

int main() {
    char str[] = "PARTITION";
    int n = strlen(str);
    int i = 0;
    int j = n-1;

    printf("%s\n", str);
    partition(str, 0, n-1, &i, &j);
    printf("%s\n", str);
    quicksort(str, 0, n-1);
    printf("%s\n", str);
   
    return 0;
} 

cc quicksort.c && ./a.out
PARTITION
AIRTPTION
AIINOPRTT

Implementation in Rust

To implement quicksort, we need rand

[package]
name = "quicksort"
version = "0.1.0"
edition = "2024"

[dependencies]
rand = "0.10.0"

Functions

main ๐